centralize components: move from app/components to src/components
- Move all components from src/app/components/ to src/components/
- Follows Next.js 14 best practices for shared components
- Components are now outside app directory for better reusability
- Includes: 3DCanvas, buttons, inputs, navigation, banners, cards
- Makes components accessible via @/components/* imports
