+++ packages/dapp/src/components/3DCanvas/models/Cowry.tsx
