+++ packages/dapp/src/components/3DCanvas/models/Drum.tsx
