+++ packages/dapp/src/components/navigation/footer.tsx
